all the functions implemented and response format is correct

+++ app/src/main/kotlin/org/dopc/ClientServerConfig/VenueSauceData/JsonStructureDynamic.kt
@@ -33,7 +33,7 @@ data class DeliveryPricing(
 data class DistanceRange(
     val min: Int,
     val max: Int,
-    val a: Double,
+    val a: Int,
     val b: Double,
     val flag: Int?
 )
